general: unclutter App – refactor ResponsiveFeaturePanel
